+++ src/main/java/com/exacttarget/fuelsdk/ClearAttachmentsOutInterceptor.java
@@ -1,5 +1,11 @@
 package com.exacttarget.fuelsdk;
 
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import org.apache.cxf.message.Message;
 import org.apache.cxf.attachment.AttachmentSerializer;
 import org.apache.cxf.binding.soap.SoapMessage;
 import org.apache.cxf.binding.soap.interceptor.AbstractSoapInterceptor;
@@ -15,7 +21,9 @@ public class ClearAttachmentsOutInterceptor extends AbstractSoapInterceptor {
     * Class constructor.
     */
     public ClearAttachmentsOutInterceptor() {
-        super(Phase.SETUP_ENDING);
+//        super(Phase.SETUP_ENDING);
+        super(Phase.POST_LOGICAL);
+
     }
 
     /** 
@@ -25,5 +33,14 @@ public class ClearAttachmentsOutInterceptor extends AbstractSoapInterceptor {
         message.getExchange()
                .getOutMessage()
                .setContent(org.apache.cxf.attachment.AttachmentSerializer.class, null);
+        
+        Map<String, List> headers = new HashMap<String, List>();
+        try {
+            headers.put("User-Agent", Collections.singletonList(message.get("HTTP_HEADER_USER_AGENT")));
+            
+            message.put(Message.PROTOCOL_HEADERS, headers);
+        } catch (Exception ce) {
+            throw new Fault(ce);
+        }
     }
 }