+++ src/main/java/com/exacttarget/fuelsdk/ETRestConnection.java
@@ -46,6 +46,10 @@ import java.net.URL;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonParser;
+import java.util.List;
+import java.util.Map;
+import java.util.logging.Level;
+import org.apache.commons.lang.StringUtils;
 
 import org.apache.log4j.Logger;
 
@@ -225,29 +229,35 @@ public class ETRestConnection {
         Gson gson = client.getGson();
 
         logger.debug(method + " " + url);
-
+        String[] token = url.getPath().split("/");
+        String object = "";
+        if(StringUtils.isNumeric(token[token.length-1]))
+            object = token[token.length-2];
+        else 
+            object = token[token.length-1];
+        
         HttpURLConnection connection = null;
         try {
             connection = (HttpURLConnection) url.openConnection();
+            
+            connection.setRequestProperty("User-Agent", "FuelSDK-Java-v1.2.1-REST-"+method+"-"+object);
             connection.setRequestMethod(method.toString());
         } catch (ProtocolException ex) {
             throw new ETSdkException("error setting request method: " + method.toString(), ex);
         } catch (IOException ex) {
             throw new ETSdkException("error opening " + url, ex);
         }
-
+  
         switch(method) {
           case GET:
             connection.setDoInput(true);
             connection.setRequestProperty("Accept", "application/json");
-            connection.setRequestProperty("User-Agent", "FuelSDK-Java-v1.2.0");
             break;
           case POST:
           case PATCH:
           case DELETE:
             connection.setDoOutput(true);
             connection.setRequestProperty("Content-Type", "application/json");
-            connection.setRequestProperty("User-Agent", "FuelSDK-Java-v1.2.0");
             break;
           default:
             throw new ETSdkException("unsupported request method: " + method.toString());
@@ -257,6 +267,13 @@ public class ETRestConnection {
             connection.setRequestProperty("Authorization", "Bearer " + client.refreshToken());
         }
 
+        try {
+            connection.connect();
+        } catch (IOException ex) {
+            java.util.logging.Logger.getLogger(ETRestConnection.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        
+        
         if (logger.isDebugEnabled()) {
             for (String key : connection.getRequestProperties().keySet()) {
                 logger.debug(key + ": " + connection.getRequestProperty(key));
@@ -295,7 +312,7 @@ public class ETRestConnection {
         } catch (IOException ex) {
             throw new ETSdkException("error getting response code / message", ex);
         }
-
+        
         return connection;
     }